requirements.txt
numpy==1.26.4
opencv-python
PyYAML

paddlepaddle==2.6.2
paddleocr==2.7.0

ultralytics==8.3.200
onnx==1.19.0
onnxruntime==1.22.1

gpiozero
RPi.GPIO
lgpio
